+++ internal/pkg/clients/init.go
@@ -47,4 +47,11 @@ var AllClients map[string][]string = map[string][]string{
 		"opnethermind",
 		"opreth",
 	},
+	"taiko": {
+		"taikoclient",
+	},
+	"texecution": {
+		"taikogeth",
+		"taikonethermind",
+	},
 }